+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/EffectBootsWithMines.lua
@@ -102,8 +102,6 @@ function EffectBootsWithMines:OnTick(deltaMs)
         ChaosVehicle.ExitVehicle(player)
         ChaosUtils.TriggerExplosionAt(square, 3)
 
-        player:setKnockedDown(true)
-
         self.visitedTiles = {}
         self:ClearSpawnedObjects()
     end