+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/Player/EffectMoveOrGetDamage.lua
@@ -56,7 +56,6 @@ function EffectMoveOrGetDamage:OnTick(deltaMs)
     if not square then return end
 
     ChaosUtils.TriggerExplosionAt(square, 3)
-    player:setKnockedDown(true)
 
     ChaosPlayer.SayLineByColor(player, "BOOM!", ChaosPlayerChatColors.red)
 end